requirements.txt
pytorch-transformers==1.1.0
sklearn==0.0
tokenizers==0.8.1rc1
transformers==3.0.2
wandb==0.10.2
